have the mnist data / mnist project working (yet again) this time in the projects folder.
it took altogether way too long to train against MNIST data set this time. Not sure why but it was painful as heck - 1000+ minutes - and much longer than previously not sure why. Perhaps because unable to use GPU but even then I think I've used the CPU for training on this data set with much faster results in the past.
A possibly confounding factor was the use of pcd-k with k=2 rather than 1 as before, but I dont think that should result in such a massive slow down (maybe?)

so far I've confirmed steps up to and including get_codes.py. get_coords.py is currently running. Having recently rewriting to python not suprising that that, also, is a lot slower than before. Not impossibly slower but a shame to see really.

So. Need to optimize for speed again. The suggestion would be to optimize for speed using fake_txt as the example since it's a much simpler network.

I'll check in the trained weights filei, or drop it into the drop box -./projects/mnist/data/ unsupervised_mnist_weights.pkl.gz - presently.