canvas: tidy up fuzz generation with an eye to scaling and use in density maps.
Changes:

* Rather than randomly permute the curve parameters until the visible
curve fits in the desired square, choose reasonable parameters and
clip the image around the visible part.

* Generate curves at 15 different densities.  Now, rather than paste
the fuzz N times on the same place for N coincident points, paste the
image with Nth density.  It will spread further and nicer.  (values of
N > 15 will later be scaled down and painted on a second map, which
can then be re-scaled and merged with the first.  But that is not done
yet).

* The curves are less steep, with lower peaks.