canvas: split $hm global object up into $const, $waiters, $page, $state, and $timestamp
This is to expedite zooming and panning and reflecting state in the
URL.

$const members are constant throughout the page session

$page members are constant once the basic map data is loaded (point
extrema, etc)

$waiters are $.Deferred objects. hmm, might rename.

$timestamp is the old $hm.timer.checkpoint

$state contains pan and zoom state and will hold the density search
string.  It maps one to one with the URL query string.

(I'm seeing the need for, say, $var, which would hold temporary data
like density maps, but which doesn't need to be stored to recreate the
page).