canvas: first go at zoom-adaptive gaussian convolution
To zoom in, you need to proportionally increase the variance (i.e.,
the effective radius) of the kernel, and you need to take into account
off-canvas points whose influence might spill on canvas.  At high levels
of zoom, the influential off-screen points outnumber the on-screen points.
This commit DOESN'T do this, and the the zoomed images reflect this.  It
looks as if the have other troubles too.

Though the overall number of points involved does keep reducing, the
increasing radius makes the calculation slower and slower.