canvas: reveal more detail in density cloud as you zoom
Rather than keep perfect scale when zooming in (as the landscape
does), the density overlay will reduce its variance and begin to show
local detail in the density map.

There is an inverse variance constant, k, which would be divided by
scale ^ 2 to stay consistent with the zoom, or left constant to stay
consistent with the canvas size.  The obvious middle solution is to
divide by scale ^ 1 itself, and is what we do (effectively).