trigrams: Condense multiple repetitions of 1 or two byte sequences
This message:

im at the doctors tweeting wwwhhhhhhhhhhhhhhhhooooooooooaaaaaaaa!

was previously evaluated as having multiple instances of 'hhh', 'ooo',
and 'aaa' trigrams.  As these are rare in English (or any language,
for that matter), it was evaluated as foreign (-0.45).

Now 4 or more repetitions of one or two characters sequences are
treated as 3 repetitions.  So that message becomes:

im at the doctors tweeting wwwhhhoooaaa!

which evaluates as a narrow pass (0.49).

In other messages the change is even more dramatic:

"Byeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee!" goes from
-2.14 to 0.81.

Repetitions of 2 character sequences are shortened to condense
the likes of "hahahahaha".