Make open source
Don't imagine there will be anybody especially keen to  pick it up and
run with this but you never know....

As agreed with other Hashmapd founders today making all this stuff open
source.

If you find you actually want to use this and have a use case that
requires something more relaxed that AGPLv3 license, let us know. We'd
probably be happy to help relax restrictions!